+++ packages/@jsii/go-runtime-test/project/tools.go
@@ -1,3 +1,4 @@
+//go:build tools
 // +build tools
 
 // Package tools contains the necessary statements to ensure tool dependencies