+++ packages/@jsii/go-runtime/jsii-runtime-go/internal/kernel/conversions.go
@@ -71,7 +71,7 @@ func (c *Client) castAndSetToPtr(ptr reflect.Value, data reflect.Value) {
 
 		// If return data is jsii object references, add to objects table.
 		if err := c.Types().InitJsiiProxy(ptr); err == nil {
-			if err = c.RegisterInstance(ptr, ref.InstanceID); err != nil {
+			if err = c.RegisterInstance(ptr, ref); err != nil {
 				panic(err)
 			}
 		} else {
@@ -157,7 +157,7 @@ func (c *Client) CastPtrToRef(dataVal reflect.Value) interface{} {
 
 	case reflect.Interface, reflect.Ptr:
 		if valref, valHasRef := c.FindObjectRef(dataVal); valHasRef {
-			return api.ObjectRef{InstanceID: valref}
+			return valref
 		}
 
 		// In case we got a pointer to a map, slice, enum, ...