+++ packages/jsii-rosetta/test/translations/classes/class_implementing_jsii_interface.java
@@ -1,0 +1,5 @@
+public class MyClass implements IResolvable {
+    public Object resolve() {
+        return 42;
+    }
+}