+++ packages/jsii/lib/assembler.ts
@@ -1,7 +1,7 @@
 import * as spec from '@jsii/spec';
 import { PackageJson } from '@jsii/spec';
 import * as Case from 'case';
-import * as colors from 'colors/safe';
+import * as chalk from 'chalk';
 import * as crypto from 'crypto';
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 import deepEqual = require('deep-equal');
@@ -183,7 +183,7 @@ export class Assembler implements Emitter {
 
       if (LOG.isTraceEnabled()) {
         LOG.trace(
-          `Processing source file: ${colors.blue(
+          `Processing source file: ${chalk.blue(
             path.relative(this.projectInfo.projectRoot, sourceFile.fileName),
           )}`,
         );
@@ -291,7 +291,7 @@ export class Assembler implements Emitter {
     const validationResult = await validator.emit();
     if (!validationResult.emitSkipped) {
       const assemblyPath = path.join(this.projectInfo.projectRoot, '.jsii');
-      LOG.trace(`Emitting assembly: ${colors.blue(assemblyPath)}`);
+      LOG.trace(`Emitting assembly: ${chalk.blue(assemblyPath)}`);
       await fs.writeJson(assemblyPath, _fingerprint(assembly), {
         encoding: 'utf8',
         spaces: 2,
@@ -525,17 +525,14 @@ export class Assembler implements Emitter {
       return `unknown.${typeName}`;
     }
 
-    // If the symbol comes from the current assembly or an assembly whose
-    // submodules we've already spidered, look up in the tables we are currently building
-    if (pkg.name === this.projectInfo.name) {
-      const submodule = this._submoduleMap.get(sym);
-      if (submodule != null) {
-        const submoduleNs =
-          this._submodules.get(submodule)!.fqnResolutionPrefix;
-        return `${submoduleNs}.${typeName}`;
-      }
-
-      return `${this.projectInfo.name}.${typeName}`;
+    // If the symbol comes from an assembly whose submodules we've already
+    // spidered (or from the current assembly), look up there. This relies
+    // on an entry-point import of the library having been done first
+    // (`import * as x from 'module-root';`)
+    const submodule = this._submoduleMap.get(sym);
+    if (submodule != null) {
+      const submoduleNs = this._submodules.get(submodule)!.fqnResolutionPrefix;
+      return `${submoduleNs}.${typeName}`;
     }
 
     // This is the fallback: in case we can't find a symbolId for the given
@@ -544,13 +541,24 @@ export class Assembler implements Emitter {
     // most likely won't be using submodules so this legacy guess will be correct.
     const fallbackFqn = `${pkg.name}.${typeName}`;
 
+    // If the type is coming from the current module, we won't find it in a dependency
+    if (pkg.name === this.projectInfo.name) {
+      return fallbackFqn;
+    }
+
     // Otherwise look up the symbol identifier in the dependency assemblies
+    // This is now the preferred mechanism but we can't do this as the only mechanism,
+    // as we may still have compile against very old assemblies that don't have a
+    // symbol identifier table at all.
     const dep = this.projectInfo.dependencyClosure.find(
       (d) => d.name === pkg.name,
     );
     if (!dep) {
       this._diagnostics.push(
-        JsiiDiagnostic.JSII_9000_UNKNOWN_MODULE.createDetached(pkg.name),
+        JsiiDiagnostic.JSII_9000_UNKNOWN_MODULE.create(
+          typeAnnotationNode,
+          pkg.name,
+        ),
       );
       return fallbackFqn;
     }
@@ -679,21 +687,17 @@ export class Assembler implements Emitter {
       return;
     }
 
-    // Normalize the path so the correct separator is in use (Looking at you, Windows)
-    resolution.resolvedModule.resolvedFileName = path.normalize(
-      resolution.resolvedModule.resolvedFileName,
-    );
     if (
       // We're not looking into a dependency's namespace exports, and the resolution says it's external
       (packageRoot === this.projectInfo.projectRoot &&
         resolution.resolvedModule.isExternalLibraryImport) ||
       // Or the module resolves outside of the current dependency's tree entirely
-      !resolution.resolvedModule.resolvedFileName.startsWith(packageRoot) ||
+      !isUnder(resolution.resolvedModule.resolvedFileName, packageRoot) ||
       // Or the module is under one the current dependency's node_modules subtree
       resolution.resolvedModule.resolvedFileName
-        .split(path.sep)
+        .split('/') // Separator is always '/', even on Windows
         .filter((entry) => entry === 'node_modules').length !==
-        packageRoot.split(path.sep).filter((entry) => entry === 'node_modules')
+        packageRoot.split('/').filter((entry) => entry === 'node_modules')
           .length
     ) {
       // External re-exports are "pure-javascript" sugar; they need not be
@@ -913,7 +917,7 @@ export class Assembler implements Emitter {
 
       if (LOG.isTraceEnabled()) {
         LOG.trace(
-          `Entering submodule: ${colors.cyan(
+          `Entering submodule: ${chalk.cyan(
             [...context.namespace, symbol.name].join('.'),
           )}`,
         );
@@ -928,7 +932,7 @@ export class Assembler implements Emitter {
 
       if (LOG.isTraceEnabled()) {
         LOG.trace(
-          `Leaving submodule: ${colors.cyan(
+          `Leaving submodule: ${chalk.cyan(
             [...context.namespace, symbol.name].join('.'),
           )}`,
         );
@@ -990,7 +994,7 @@ export class Assembler implements Emitter {
 
       if (LOG.isTraceEnabled()) {
         LOG.trace(
-          `Entering namespace: ${colors.cyan(
+          `Entering namespace: ${chalk.cyan(
             [...context.namespace, name].join('.'),
           )}`,
         );
@@ -1009,7 +1013,7 @@ export class Assembler implements Emitter {
 
       if (LOG.isTraceEnabled()) {
         LOG.trace(
-          `Leaving namespace:  ${colors.cyan(
+          `Leaving namespace:  ${chalk.cyan(
             [...context.namespace, name].join('.'),
           )}`,
         );
@@ -1065,7 +1069,7 @@ export class Assembler implements Emitter {
 
     if (LOG.isInfoEnabled()) {
       LOG.info(
-        `Registering JSII ${colors.magenta(jsiiType.kind)}: ${colors.green(
+        `Registering JSII ${chalk.magenta(jsiiType.kind)}: ${chalk.green(
           jsiiType.fqn,
         )}`,
       );
@@ -1223,9 +1227,9 @@ export class Assembler implements Emitter {
   ): Promise<spec.ClassType | undefined> {
     if (LOG.isTraceEnabled()) {
       LOG.trace(
-        `Processing class: ${colors.gray(
-          ctx.namespace.join('.'),
-        )}.${colors.cyan(type.symbol.name)}`,
+        `Processing class: ${chalk.gray(ctx.namespace.join('.'))}.${chalk.cyan(
+          type.symbol.name,
+        )}`,
       );
     }
 
@@ -1270,7 +1274,7 @@ export class Assembler implements Emitter {
       // erased, and identify the closest exported base class, should there be one.
       while (base && this._isPrivateOrInternal(base.symbol)) {
         LOG.debug(
-          `Base class of ${colors.green(jsiiType.fqn)} named ${colors.green(
+          `Base class of ${chalk.green(jsiiType.fqn)} named ${chalk.green(
             base.symbol.name,
           )} is not exported, erasing it...`,
         );
@@ -1684,7 +1688,7 @@ export class Assembler implements Emitter {
 
     if (_isPrivate(symbol)) {
       LOG.trace(
-        `${colors.cyan(
+        `${chalk.cyan(
           symbol.name,
         )} is marked "private", or is an unexported type declaration`,
       );
@@ -1725,7 +1729,7 @@ export class Assembler implements Emitter {
   ): Promise<spec.EnumType | undefined> {
     if (LOG.isTraceEnabled()) {
       LOG.trace(
-        `Processing enum: ${colors.gray(ctx.namespace.join('.'))}.${colors.cyan(
+        `Processing enum: ${chalk.gray(ctx.namespace.join('.'))}.${chalk.cyan(
           type.symbol.name,
         )}`,
       );
@@ -1876,9 +1880,9 @@ export class Assembler implements Emitter {
   ): Promise<spec.InterfaceType | undefined> {
     if (LOG.isTraceEnabled()) {
       LOG.trace(
-        `Processing interface: ${colors.gray(
+        `Processing interface: ${chalk.gray(
           ctx.namespace.join('.'),
-        )}.${colors.cyan(type.symbol.name)}`,
+        )}.${chalk.cyan(type.symbol.name)}`,
       );
     }
 
@@ -2116,7 +2120,7 @@ export class Assembler implements Emitter {
   ) {
     if (LOG.isTraceEnabled()) {
       LOG.trace(
-        `Processing method: ${colors.green(type.fqn)}#${colors.cyan(
+        `Processing method: ${chalk.green(type.fqn)}#${chalk.cyan(
           symbol.name,
         )}`,
       );
@@ -2237,7 +2241,7 @@ export class Assembler implements Emitter {
       ) != null
     ) {
       LOG.trace(
-        `Dropping re-declaration of ${colors.green(type.fqn)}#${colors.cyan(
+        `Dropping re-declaration of ${chalk.green(type.fqn)}#${chalk.cyan(
           method.name,
         )}`,
       );
@@ -2281,7 +2285,7 @@ export class Assembler implements Emitter {
 
     if (LOG.isTraceEnabled()) {
       LOG.trace(
-        `Processing property: ${colors.green(type.fqn)}#${colors.cyan(
+        `Processing property: ${chalk.green(type.fqn)}#${chalk.cyan(
           symbol.name,
         )}`,
       );
@@ -2369,7 +2373,7 @@ export class Assembler implements Emitter {
       ) != null
     ) {
       LOG.trace(
-        `Dropping re-declaration of ${colors.green(type.fqn)}#${colors.cyan(
+        `Dropping re-declaration of ${chalk.green(type.fqn)}#${chalk.cyan(
           property.name,
         )}`,
       );
@@ -2384,7 +2388,7 @@ export class Assembler implements Emitter {
     ctx: EmitContext,
   ): Promise<spec.Parameter> {
     if (LOG.isTraceEnabled()) {
-      LOG.trace(`Processing parameter: ${colors.cyan(paramSymbol.name)}`);
+      LOG.trace(`Processing parameter: ${chalk.cyan(paramSymbol.name)}`);
     }
     const paramDeclaration =
       paramSymbol.valueDeclaration as ts.ParameterDeclaration;
@@ -2589,11 +2593,6 @@ export class Assembler implements Emitter {
       }
       // Not a primitive type!
       return undefined;
-
-      function isUnder(file: string, dir: string): boolean {
-        const relative = path.relative(dir, file);
-        return !relative.startsWith(path.sep) && !relative.startsWith('..');
-      }
     }
 
     async function _unionType(this: Assembler): Promise<spec.OptionalValue> {
@@ -3476,4 +3475,9 @@ function getSymbolFromDeclaration(
 ): ts.Symbol | undefined {
   const name = ts.getNameOfDeclaration(decl);
   return name ? typeChecker.getSymbolAtLocation(name) : undefined;
+}
+
+function isUnder(file: string, dir: string): boolean {
+  const relative = path.relative(dir, file);
+  return !relative.startsWith(path.sep) && !relative.startsWith('..');
 }